Prevent running the tests in BinaryLogClientIntegrationTest twice
